Switch to client credentials; make all model fields nullable
- Changed Salesforce auth to use ClientCredentialsAsync (OAuth 2.0 client credentials flow).
- Generated model classes are now partial for easier extension.
- All model fields are now nullable by default, not just select types.
- Mapped double_sf and percent_sf to decimal in SfTypeConverter for better precision.

+++ src/NetCoreForce.ModelGenerator/SfTypeConverter.cs
@@ -31,7 +31,7 @@ namespace NetCoreForce.ModelGenerator
                     type = typeof(DateTime);
                     break;
                 case FieldType.double_sf:
-                    type = typeof(double);
+                    type = typeof(decimal);
                     break;
                 case FieldType.int_sf:
                     type = typeof(int);
@@ -97,7 +97,7 @@ namespace NetCoreForce.ModelGenerator
                     type = typeof(string);
                     break;
                 case FieldType.percent_sf:
-                    type = typeof(double);
+                    type = typeof(decimal);
                     break;
                 case FieldType.phone_sf:
                     type = typeof(string);
